+++ CHANGELOG.md
@@ -9,7 +9,9 @@ and [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ### Added
 
-- store HTTP responses in ~/.cache/jokeyrhyme-dotfiles/
+- cache HTTP responses in ~/.cache/jokeyrhyme-dotfiles/
+
+- use cached HTTP response if less than 15 minutes old
 
 ## [0.32.0] - 2019-05-28